move storypanel out of tab dir
