+++ web/src/beta/features/Editor/StoryPanel/index.tsx
